Update/rename usermode_rotary_brightness_color, to be used as an example of more robust parsing Usermod Settings values
